fix: tech debt — widen sourceRefs to uint32, add NamedCallable interface
Task 1.1: sourceRefs []uint16 → []uint32 across native_template.go,
edit_plan.go, peephole.go, and all test files. Raises the source table
limit from 65K to 4B entries. The old uint16 had a panic guard but
couldn't represent large templates; the widening removes the constraint.

Task 5.1: Add NamedCallable interface (Name/Doc) to machine/closure.go,
embedded in the Closure interface. Compile-time checks for
MachineClosure and CaseLambdaClosure. Replaced anonymous interfaces in
prim_reflection.go with the named type.

Also: mark Tasks 1.2, 1.3, 4.2 as already implemented in the tech debt
plan; add new tasks 1.4, 5.5, 6.4, 8.4 from full-codebase reassessment;
update TODO.md with consolidated tech debt section.

+++ machine/edit_plan_test.go
@@ -27,7 +27,7 @@ import (
 func TestEditPlan_EmptyPlan(t *testing.T) {
 	tpl := NewEmptyNativeTemplate()
 	tpl.code = []Instruction{{Op: OpLoadVoid}, {Op: OpPush}}
-	tpl.sourceRefs = []uint16{1, 2}
+	tpl.sourceRefs = []uint32{1, 2}
 
 	plan := NewEditPlan(tpl)
 	delta := plan.Apply()
@@ -39,7 +39,7 @@ func TestEditPlan_EmptyPlan(t *testing.T) {
 func TestEditPlan_HasEdits(t *testing.T) {
 	tpl := NewEmptyNativeTemplate()
 	tpl.code = []Instruction{{Op: OpLoadVoid}}
-	tpl.sourceRefs = []uint16{0}
+	tpl.sourceRefs = []uint32{0}
 
 	plan := NewEditPlan(tpl)
 	qt.Assert(t, plan.HasEdits(), qt.IsFalse)
@@ -58,7 +58,7 @@ func TestEditPlan_SingleDelete(t *testing.T) {
 		{Op: OpLoadLiteral, Arg: 1},
 		{Op: OpPush},
 	}
-	tpl.sourceRefs = []uint16{1, 2, 3, 4}
+	tpl.sourceRefs = []uint32{1, 2, 3, 4}
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(1, 2)
@@ -66,7 +66,7 @@ func TestEditPlan_SingleDelete(t *testing.T) {
 
 	qt.Assert(t, delta, qt.Equals, -1)
 	qt.Assert(t, opcodes(tpl), qt.DeepEquals, []OpCode{OpStoreGlobal, OpLoadLiteral, OpPush})
-	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint16{1, 3, 4})
+	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint32{1, 3, 4})
 }
 
 func TestEditPlan_DeleteRange(t *testing.T) {
@@ -78,14 +78,14 @@ func TestEditPlan_DeleteRange(t *testing.T) {
 		{Op: OpLoadLiteral, Arg: 1}, // 3: delete
 		{Op: OpApply},               // 4: keep
 	}
-	tpl.sourceRefs = []uint16{1, 2, 3, 4, 5}
+	tpl.sourceRefs = []uint32{1, 2, 3, 4, 5}
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(1, 4)
 	plan.Apply()
 
 	qt.Assert(t, opcodes(tpl), qt.DeepEquals, []OpCode{OpStoreGlobal, OpApply})
-	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint16{1, 5})
+	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint32{1, 5})
 }
 
 func TestEditPlan_MultipleDeletes(t *testing.T) {
@@ -97,7 +97,7 @@ func TestEditPlan_MultipleDeletes(t *testing.T) {
 		{Op: OpLoadVoid},            // 3: delete
 		{Op: OpApply},               // 4: keep
 	}
-	tpl.sourceRefs = []uint16{1, 2, 3, 4, 5}
+	tpl.sourceRefs = []uint32{1, 2, 3, 4, 5}
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(1, 2)
@@ -105,7 +105,7 @@ func TestEditPlan_MultipleDeletes(t *testing.T) {
 	plan.Apply()
 
 	qt.Assert(t, opcodes(tpl), qt.DeepEquals, []OpCode{OpLoadLiteral, OpPush, OpApply})
-	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint16{1, 3, 5})
+	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint32{1, 3, 5})
 }
 
 // --- Replace ---
@@ -125,7 +125,7 @@ func TestEditPlan_SingleReplace(t *testing.T) {
 		{Op: OpApply},                    // 8
 		{Op: OpPush},                     // 9: after the call
 	}
-	tpl.sourceRefs = []uint16{10, 10, 10, 10, 10, 10, 10, 10, 10, 11}
+	tpl.sourceRefs = []uint32{10, 10, 10, 10, 10, 10, 10, 10, 10, 11}
 	tpl.literals = MultipleValues{values.NewSymbol("+"), values.NewInteger(3), values.NewInteger(4)}
 
 	plan := NewEditPlan(tpl)
@@ -137,7 +137,7 @@ func TestEditPlan_SingleReplace(t *testing.T) {
 
 	qt.Assert(t, opcodes(tpl), qt.DeepEquals, []OpCode{OpLoadLiteral, OpPush})
 	qt.Assert(t, tpl.code[0].Arg, qt.Equals, litIdx)
-	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint16{10, 11})
+	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint32{10, 11})
 }
 
 func TestEditPlan_ReplaceWithLarger(t *testing.T) {
@@ -147,7 +147,7 @@ func TestEditPlan_ReplaceWithLarger(t *testing.T) {
 		{Op: OpLoadVoid}, // 0: replace with 3 instrs
 		{Op: OpPush},     // 1: keep
 	}
-	tpl.sourceRefs = []uint16{1, 2}
+	tpl.sourceRefs = []uint32{1, 2}
 
 	plan := NewEditPlan(tpl)
 	plan.Replace(0, 1, []Instruction{
@@ -158,7 +158,7 @@ func TestEditPlan_ReplaceWithLarger(t *testing.T) {
 	plan.Apply()
 
 	qt.Assert(t, opcodes(tpl), qt.DeepEquals, []OpCode{OpLoadLiteral, OpPush, OpPop, OpPush})
-	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint16{5, 5, 5, 2})
+	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint32{5, 5, 5, 2})
 }
 
 // --- Insert ---
@@ -169,7 +169,7 @@ func TestEditPlan_Insert(t *testing.T) {
 		{Op: OpLoadVoid}, // 0
 		{Op: OpPush},     // 1
 	}
-	tpl.sourceRefs = []uint16{1, 2}
+	tpl.sourceRefs = []uint32{1, 2}
 
 	plan := NewEditPlan(tpl)
 	plan.Insert(1, []Instruction{
@@ -178,7 +178,7 @@ func TestEditPlan_Insert(t *testing.T) {
 	plan.Apply()
 
 	qt.Assert(t, opcodes(tpl), qt.DeepEquals, []OpCode{OpLoadVoid, OpLoadLiteral, OpPush})
-	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint16{1, 3, 2})
+	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint32{1, 3, 2})
 }
 
 // --- Branch Fixup ---
@@ -193,7 +193,7 @@ func TestEditPlan_BranchShrinks(t *testing.T) {
 		{Op: OpLoadLiteral, Arg: 1}, // 3
 		{Op: OpApply},               // 4
 	}
-	tpl.sourceRefs = make([]uint16, 5)
+	tpl.sourceRefs = make([]uint32, 5)
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(2, 3)
@@ -211,7 +211,7 @@ func TestEditPlan_BranchGrows(t *testing.T) {
 		{Op: OpPush},           // 1
 		{Op: OpApply},          // 2
 	}
-	tpl.sourceRefs = make([]uint16, 3)
+	tpl.sourceRefs = make([]uint32, 3)
 
 	plan := NewEditPlan(tpl)
 	plan.Insert(1, []Instruction{
@@ -237,7 +237,7 @@ func TestEditPlan_BranchAcrossReplace(t *testing.T) {
 		{Op: OpPush},           // 5
 		{Op: OpApply},          // 6
 	}
-	tpl.sourceRefs = make([]uint16, 7)
+	tpl.sourceRefs = make([]uint32, 7)
 
 	plan := NewEditPlan(tpl)
 	plan.Replace(2, 5, []Instruction{
@@ -258,7 +258,7 @@ func TestEditPlan_SaveContinuationSentinel(t *testing.T) {
 		{Op: OpLoadLiteral, Arg: 0},      // 2
 		{Op: OpApply},                    // 3
 	}
-	tpl.sourceRefs = make([]uint16, 4)
+	tpl.sourceRefs = make([]uint32, 4)
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(1, 2)
@@ -277,7 +277,7 @@ func TestEditPlan_BranchBeforeReplace_NotSpanning(t *testing.T) {
 		{Op: OpLoadVoid},        // 2: delete
 		{Op: OpApply},           // 3
 	}
-	tpl.sourceRefs = make([]uint16, 4)
+	tpl.sourceRefs = make([]uint32, 4)
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(2, 3)
@@ -292,7 +292,7 @@ func TestEditPlan_BranchBeforeReplace_NotSpanning(t *testing.T) {
 func TestEditPlan_AddLiteral_New(t *testing.T) {
 	tpl := NewEmptyNativeTemplate()
 	tpl.code = []Instruction{{Op: OpLoadVoid}}
-	tpl.sourceRefs = []uint16{0}
+	tpl.sourceRefs = []uint32{0}
 	tpl.MaybeAppendLiteral(values.NewInteger(1))
 
 	plan := NewEditPlan(tpl)
@@ -305,7 +305,7 @@ func TestEditPlan_AddLiteral_New(t *testing.T) {
 func TestEditPlan_AddLiteral_Dedup(t *testing.T) {
 	tpl := NewEmptyNativeTemplate()
 	tpl.code = []Instruction{{Op: OpLoadVoid}}
-	tpl.sourceRefs = []uint16{0}
+	tpl.sourceRefs = []uint32{0}
 	tpl.MaybeAppendLiteral(values.NewInteger(42))
 
 	plan := NewEditPlan(tpl)
@@ -328,7 +328,7 @@ func TestEditPlan_SideTableGC(t *testing.T) {
 		{Op: OpComplex, Arg: 1}, // 1: references op1 — will be deleted
 		{Op: OpComplex, Arg: 2}, // 2: references op2
 	}
-	tpl.sourceRefs = []uint16{0, 0, 0}
+	tpl.sourceRefs = []uint32{0, 0, 0}
 	tpl.sideTable = []InlinedOperation{op0, op1, op2}
 
 	plan := NewEditPlan(tpl)
@@ -352,7 +352,7 @@ func TestEditPlan_SideTableGC_NoneUnreferenced(t *testing.T) {
 		{Op: OpLoadVoid}, // delete this (doesn't reference sideTable)
 		{Op: OpLoadLiteral, Arg: 0},
 	}
-	tpl.sourceRefs = []uint16{0, 0, 0}
+	tpl.sourceRefs = []uint32{0, 0, 0}
 	tpl.sideTable = []InlinedOperation{op0}
 
 	plan := NewEditPlan(tpl)
@@ -371,7 +371,7 @@ func TestEditPlan_OverlappingEdits_Panics(t *testing.T) {
 		{Op: OpPush},
 		{Op: OpApply},
 	}
-	tpl.sourceRefs = make([]uint16, 3)
+	tpl.sourceRefs = make([]uint32, 3)
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(0, 2)
@@ -383,7 +383,7 @@ func TestEditPlan_OverlappingEdits_Panics(t *testing.T) {
 func TestEditPlan_OutOfBounds_Panics(t *testing.T) {
 	tpl := NewEmptyNativeTemplate()
 	tpl.code = []Instruction{{Op: OpLoadVoid}}
-	tpl.sourceRefs = []uint16{0}
+	tpl.sourceRefs = []uint32{0}
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(0, 5) // end > len(code)
@@ -401,7 +401,7 @@ func TestEditPlan_MixedEdits(t *testing.T) {
 		{Op: OpLoadLiteral, Arg: 0}, // 2: replace with 2 instrs
 		{Op: OpApply},               // 3: keep
 	}
-	tpl.sourceRefs = []uint16{1, 2, 3, 4}
+	tpl.sourceRefs = []uint32{1, 2, 3, 4}
 
 	plan := NewEditPlan(tpl)
 	plan.Delete(0, 1)
@@ -414,7 +414,7 @@ func TestEditPlan_MixedEdits(t *testing.T) {
 	// New: [Push, LoadGlobal, Pop, Apply]
 	qt.Assert(t, delta, qt.Equals, 0) // -1 + 1 = 0 net
 	qt.Assert(t, opcodes(tpl), qt.DeepEquals, []OpCode{OpPush, OpLoadGlobal, OpPop, OpApply})
-	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint16{2, 7, 7, 4})
+	qt.Assert(t, tpl.sourceRefs, qt.DeepEquals, []uint32{2, 7, 7, 4})
 }
 
 // --- buildEditRemap ---
